fix(workflows): deliver custom-agent skills/MCP snapshots at step spawn
The workflow engine predates custom agents gaining skills and MCP tool
servers, so a step assigned a custom agent silently spawned without them.
Extract the draft spawn path's snapshot resolution (assignment order,
dangling-id drop, undeliverable-server filter) into a shared
snapshotAgentDeliverables helper and pass its result from both spawn sites.
